Change vis_uvw_fp type to double

+++ src/hpg.hpp
@@ -127,7 +127,7 @@ using grid_value_fp = double;
 /** grid scale floating point type */
 using grid_scale_fp = float;
 /** visibility (U, V, W) coordinate floating point type */
-using vis_uvw_fp = float;
+using vis_uvw_fp = double;
 /** visibility weight floating point type */
 using vis_weight_fp = float;
 /** visibility frequency floating point type */
